Move ObjLoader struct and AssetLoader impl
